Parse Short as unsiged so it doesnt crash with big files

+++ src/main/java/no/myke/parser/ModelObject.java
@@ -6,7 +6,7 @@ package no.myke.parser;
 public class ModelObject {
     private String name;
     public float[] vertices;
-    public short[] polygons;
+    public int[] polygons;
     public float[] textureCoordinates;
 
     public ModelObject(String name) {
